(CI) don't try to deploy shared object library when linking against musl

+++ ci/before_deploy.sh
@@ -21,7 +21,9 @@ main() {
 
     cp target/$TARGET/release/cres $stage/
     cp target/$TARGET/release/libcres.a $stage/
-    cp target/$TARGET/release/libcres.so $stage/
+    if [ -f target/$TARGET/release/libcres.so ]; then
+        cp target/$TARGET/release/libcres.so $stage/
+    fi
 
     cd $stage
     tar czf $src/$CRATE_NAME-$TRAVIS_TAG-$TARGET.tar.gz *
